PIC-2929 split out integration tests correctly
